# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vendors/apply": {
      "post": {
        "operationId": "VendorsController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      }
    },
    "/vendor/products": {
      "get": {
        "operationId": "VendorsController_listProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      },
      "post": {
        "operationId": "VendorsController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      }
    },
    "/vendor/products/{id}": {
      "patch": {
        "operationId": "VendorsController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      }
    },
    "/vendor/products/{id}/variants": {
      "post": {
        "operationId": "VendorsController_addVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      }
    },
    "/vendor/products/{id}/variants/{variantId}": {
      "patch": {
        "operationId": "VendorsController_updateVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      }
    },
    "/vendor/products/{id}/images": {
      "post": {
        "operationId": "VendorsController_addImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProductImagesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "vendor"
        ]
      }
    },
...
    "/returns": {
      "post": {
        "operationId": "ReturnsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      },
      "get": {
        "operationId": "ReturnsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
    "/returns/{id}/approve": {
      "patch": {
        "operationId": "ReturnsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReturnStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
    "/returns/{id}/reject": {
      "patch": {
        "operationId": "ReturnsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReturnStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
    "/returns/{id}/receive": {
      "patch": {
        "operationId": "ReturnsController_receive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReturnStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
    "/returns/{id}/refund": {
      "patch": {
        "operationId": "ReturnsController_refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReturnStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
    "/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ReviewsController_list",
        "parameters": [
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
    "/promotions": {
      "post": {
        "operationId": "PromotionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "promotions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PromotionsController_list",
        "parameters": [
          {
            "name": "vendorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter active promotions at current time",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "promotions"
        ]
      }
    },
    "/promotions/validate": {
      "post": {
        "operationId": "PromotionsController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "promotions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Eco System Order API",
    "description": "API documentation for Eco System Order App",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApplyVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "example": "US"
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "commissionPct": {
            "type": "number",
            "default": 10
          }
        },
        "required": [
          "name",
          "country",
          "currency"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "categoryId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ]
          }
        }
      },
      "VariantOptionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "value"
        ]
      },
      "CreateVariantDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantOptionDto"
            }
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "stockQty": {
            "type": "number"
          },
          "weightG": {
            "type": "number"
          }
        },
        "required": [
          "sku",
          "options",
          "price",
          "currency",
          "stockQty"
        ]
      },
      "UpdateVariantDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantOptionDto"
            }
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "stockQty": {
            "type": "number"
          },
          "weightG": {
            "type": "number"
          }
        }
      },
      "ProductImageInputDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "url"
        ]
      },
      "AddProductImagesDto": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageInputDto"
            }
          }
        },
        "required": [
          "images"
        ]
      },
...
      "CreateReturnDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "reason"
        ]
      },
      "UpdateReturnStatusDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string"
          }
        }
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "orderItemId",
          "rating"
        ]
      },
      "CreatePromotionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ECO10"
          },
          "type": {
            "type": "string",
            "enum": [
              "PERCENT",
              "FIXED"
            ],
            "example": "PERCENT"
          },
          "amount": {
            "type": "number",
            "example": 10,
            "description": "Percent(0-100) or fixed amount"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-12-31T23:59:59.999Z"
          },
          "minOrder": {
            "type": "number",
            "example": 5000
          }
        },
        "required": [
          "code",
          "type",
          "amount",
          "startsAt",
          "endsAt"
        ]
      },
      "ValidatePromotionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ECO10"
          },
          "orderTotal": {
            "type": "number",
            "example": 10000,
            "description": "Order total in smallest unit"
          },
          "vendorId": {
            "type": "string",
            "description": "Vendor id to validate vendor-specific promotions"
          }
        },
        "required": [
          "code",
          "orderTotal"
        ]
      },
...
    }
  }
}